plugins/tag-suggestion/pyproject.toml
[project]
name = "plugin-tag-suggestion"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    { name = "Michael", email = "80940180+michael-freling@users.noreply.github.com" },
]
dependencies = [
    "accelerate==1.1.0",
    "aiohappyeyeballs==2.4.3",
    "aiohttp==3.10.10",
    "aiosignal==1.3.1",
    "attrs==24.2.0",
    "certifi==2024.8.30",
    "charset-normalizer==3.4.0",
    "click==8.1.8",
    "datasets==3.1.0",
    "dill==0.3.8",
    "evaluate==0.4.3",
    "filelock==3.16.1",
    "frozenlist==1.5.0",
    "fsspec==2024.9.0",
    "grpcio==1.68.1",
    "huggingface-hub==0.26.2",
    "idna==3.10",
    "Jinja2==3.1.4",
    "joblib==1.4.2",
    "MarkupSafe==3.0.2",
    "mpmath==1.3.0",
    "multidict==6.1.0",
    "multiprocess==0.70.16",
    "networkx==3.4.2",
    "numpy==2.1.2",
    "nvidia-cublas-cu12==12.1.3.1",
    "nvidia-cuda-cupti-cu12==12.1.105",
    "nvidia-cuda-nvrtc-cu12==12.1.105",
    "nvidia-cuda-runtime-cu12==12.1.105",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.0.2.54",
    "nvidia-curand-cu12==10.3.2.106",
    "nvidia-cusolver-cu12==11.4.5.107",
    "nvidia-cusparse-cu12==12.1.0.106",
    "nvidia-nccl-cu12==2.20.5",
    "nvidia-nvjitlink-cu12==12.4.127",
    "nvidia-nvtx-cu12==12.1.105",
    "packaging==24.1",
    "pandas==2.2.3",
    "peewee==3.17.8",
    "pillow==11.0.0",
    "propcache==0.2.0",
    "protobuf==5.28.3",
    "psutil==6.1.0",
    "pyarrow==18.0.0",
    "python-dateutil==2.9.0.post0",
    "python-xmp-toolkit==2.0.2",
    "pytz==2024.2",
    "PyYAML==6.0.2",
    "regex==2024.9.11",
    "requests==2.32.3",
    "safetensors==0.4.5",
    "scikit-learn==1.5.2",
    "scipy==1.14.1",
    "setuptools==75.3.0",
    "six==1.16.0",
    "sympy==1.13.1",
    "tensorboardX==2.6.2.2",
    "threadpoolctl==3.5.0",
    "tokenizers==0.20.1",
    "torch==2.4.1",
    "torch-directml==0.2.5.dev240914",
    "torchvision==0.19.1",
    "tqdm==4.66.6",
    "transformers==4.46.1",
    "triton==3.0.0",
    "typing-extensions==4.12.2",
    "tzdata==2024.2",
    "urllib3==2.2.3",
    "xxhash==3.5.0",
    "yarl==1.17.1",
    "grpc-interceptor>=0.15.4",
]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "MIT" }

[dependency-groups]
# https://pdm-project.org/en/stable/usage/advanced/#use-pdm-to-manage-a-monorepo
dev = ["-e file:///${PROJECT_ROOT}/../plugins-protos/gen/python"]

[tool.pdm]
distribution = false

[tool.pdm.scripts]
_.env = { DATA_DIR = 'source', PREPROCESSED_DATA_DIR = "datasets", MODEL_DIR = "model" }
preprocess.composite = [
    "rm -rf ${PREPROCESSED_DATA_DIR}",
    "python src/main.py preprocess ${DATA_DIR} ${PREPROCESSED_DATA_DIR}",
    "zip -rq ${PREPROCESSED_DATA_DIR}.zip ${PREPROCESSED_DATA_DIR}",
]
train.composite = [
    "rm -rf ${MODEL_DIR}",
    "python src/main.py train ${PREPROCESSED_DATA_DIR} ${MODEL_DIR}",
]
# _.env = { PYTHON_PATH = "../plugins-protos/gen/python/tag_suggestion" }
server.cmd = "python src/main.py server ${MODEL_DIR}"
client.cmd = "python src/grpc_client.py {args}"
predict.cmd = "python src/main.py predict ${MODEL_DIR} {args}"
